chore: refactor app delegate based on PR feedback
