+++ src/xcode/ENA/ENA/Source/AppDelegate.swift
@@ -114,7 +114,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 	#endif
 
 	private var exposureDetection: ExposureDetection?
-	private var exposureSubmissionService: ENAExposureSubmissionService?
+	private(set) var exposureSubmissionService: ENAExposureSubmissionService?
 
 	let downloadedPackagesStore: DownloadedPackagesStore = DownloadedPackagesSQLLiteStore(fileName: "packages")
 
@@ -158,144 +158,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 	func application(_: UIApplication, didDiscardSceneSessions _: Set<UISceneSession>) {}
 }
 
-extension AppDelegate: ENATaskExecutionDelegate {
-
-	/// This method executes the background tasks needed for: a) fetching test results and b) performing exposure detection requests
-	func executeENABackgroundTask(task: BGTask, completion: @escaping ((Bool) -> Void)) {
-		executeFetchTestResults(task: task) { fetchTestResultSuccess in
-
-			// NOTE: We are currently fetching the test result first, and then execute
-			// the exposure detection check. Instead of implementing this behaviour in the completion handler,
-			// queues could be used as well. Due to time/resource constraints, we settled for this option.
-			self.executeExposureDetectionRequest(task: task) { exposureDetectionSuccess in
-				self.executeFakeRequests() {
-					completion(fetchTestResultSuccess && exposureDetectionSuccess)
-				}
-			}
-		}
-	}
-
-	/// This method executes a  test result fetch, and if it is successful, and the test result is different from the one that was previously
-	/// part of the app, a local notification is shown.
-	/// NOTE: This method will always return true.
-	private func executeFetchTestResults(task: BGTask, completion: @escaping ((Bool) -> Void)) {
-
-		exposureSubmissionService = ENAExposureSubmissionService(diagnosiskeyRetrieval: exposureManager, client: client, store: store)
-
-		if store.registrationToken != nil && store.testResultReceivedTimeStamp == nil {
-			self.exposureSubmissionService?.getTestResult { result in
-				switch result {
-				case .failure(let error):
-					logError(message: error.localizedDescription)
-				case .success(let testResult):
-					if testResult != .pending {
-						UNUserNotificationCenter.current().presentNotification(
-							title: AppStrings.LocalNotifications.testResultsTitle,
-							body: AppStrings.LocalNotifications.testResultsBody,
-							identifier: task.identifier
-						)
-					}
-				}
-
-				completion(true)
-			}
-		} else {
-			completion(true)
-		}
-
-	}
-
-	/// This method performs a check for the current exposure detection state. Only if the risk level has changed compared to the
-	/// previous state, a local notification is shown.
-	/// NOTE: This method will always return true.
-	private func executeExposureDetectionRequest(task: BGTask, completion: @escaping ((Bool) -> Void)) {
-
-		let detectionMode = DetectionMode.fromBackgroundStatus()
-		riskProvider.configuration.detectionMode = detectionMode
-
-		riskProvider.requestRisk(userInitiated: false) { risk in
-			// present a notification if the risk score has increased.
-			if let risk = risk,
-				risk.riskLevelHasChanged {
-				UNUserNotificationCenter.current().presentNotification(
-					title: AppStrings.LocalNotifications.detectExposureTitle,
-					body: AppStrings.LocalNotifications.detectExposureBody,
-					identifier: task.identifier
-				)
-			}
-			completion(true)
-		}
-	}
-
-	// MARK: - Plausible deniability.
-
-	var minHoursToNextBackgroundExecution: Double { 4 }
-	var maxHoursToNextBackgroundExecution: Double { 12 }
-	var numberOfDaysToRunPlaybook: Double { 16 }
-	var minNumberOfSequentialPlaybooks: Int { 1 }
-	var maxNumberOfSequentialPlaybooks: Int { 4 }
-	var minDelayBetweenSequentialPlaybooks: Int { 5 } // seconds
-	var maxDelayBetweenSequentialPlaybooks: Int { 10 } // seconds
-
-	/// Trigger a fake playbook to enable plausible deniability.
-	private func executeFakeRequests(_ completion: (() -> Void)? = nil) {
-		guard store.isAllowedToSubmitDiagnosisKeys else {
-			completion?()
-			return
-		}
-
-		// Initialize firstPlaybookExecution date during the first run regardless of actual execution.
-		if store.firstPlaybookExecution == nil {
-			store.firstPlaybookExecution = Date()
-		}
-
-		// Time interval until we want to resend a fake request from the background.
-		let offset = Double.random(in: minHoursToNextBackgroundExecution...maxHoursToNextBackgroundExecution) * 60
-		let now = Date()
-
-		if
-			let firstPlaybookExecution = store.firstPlaybookExecution,
-			firstPlaybookExecution.addingTimeInterval(numberOfDaysToRunPlaybook * 86_400) > now,
-			store.lastBackgroundFakeRequest.addingTimeInterval(offset) > now
-		{
-			sendFakeRequest {
-				self.store.lastBackgroundFakeRequest = now
-				completion?()
-			}
-		} else {
-			completion?()
-		}
-	}
-
-	func executeFakeRequestOnAppLaunch() {
-		// Execute a fake request 1 in 100 times while we are running in foreground.
-		guard
-			UIApplication.shared.applicationState == .active,
-			Int.random(in: 0...100) == 6
-		else { return }
-
-		sendFakeRequest()
-	}
-
-	/// Triggers one or more fake requests over a time interval of multiple seconds.
-	/// - Parameters:
-	///   - completion: called after all requests were triggered. Currently, only required when running in background mode to avoid terminating before the requests were made.
-	private func sendFakeRequest(_ completion: (() -> Void)? = nil) {
-		let service = exposureSubmissionService ?? ENAExposureSubmissionService(diagnosiskeyRetrieval: exposureManager, client: client, store: store)
-		let group = DispatchGroup()
-
-		for i in 0..<Int.random(in: minNumberOfSequentialPlaybooks...maxNumberOfSequentialPlaybooks) {
-			let delay = Int.random(in: minDelayBetweenSequentialPlaybooks...maxDelayBetweenSequentialPlaybooks)
-			group.enter()
-			DispatchQueue.global().asyncAfter(deadline: .now() + .seconds(i * delay)) {
-				service.fakeRequest()
-				group.leave()
-			}
-		}
-
-		// Wait for all fake request to finish and call completion handler.
-		group.notify(queue: .global()) {
-			completion?()
-		}
-	}
-}
